part 16 load cart items done

+++ app/src/main/java/com/aim2u/kotlineatitv2client/Database/CartDAO.kt
@@ -14,7 +14,7 @@ interface CartDAO {
     fun countItemInCart(uid:String):Single<Int>
 
     @Query("SELECT SUM(foodQuantity*foodPrice)+(foodExtraPrice*foodQuantity) FROM Cart WHERE uid=:uid")
-    fun sumPrice(uid:String):Single<Long>
+    fun sumPrice(uid:String):Single<Double>
 
     @Query("SELECT * FROM Cart WHERE foodId=:foodId AND uid=:uid")
     fun getItemInCart(foodId:String, uid:String): Single<CartItem>
